Add a file reader.
This reader will read lines from a file, periodically checking if it has been rotated or replaced.  If a time format is specified, it will assume that each line starts with a matching timestamp, discard lines that don't match, and remove the timestamp before passing the data on to the parser.